fix: 将RETAIN_TOKN从整型改为浮点型以支持小数输入

+++ models/sparse_qwen3vl/score.py
@@ -4,7 +4,7 @@ import os
 VERSION = os.getenv("USE_VERSION", "1_0")
 V2_0 = VERSION == "2_0"
 
-RETAIN_TOKN = int(os.getenv("RETAIN_TOKN", "0.58"))
+RETAIN_TOKN = float(os.getenv("RETAIN_TOKN", "0.58"))
 
 layer_dict = {5: 0, 10: 1, 20: 2}
